prioritize pointclouds and meshes equally
Pointclouds previously superseded any clicks on a mesh when there was overlap. Clicking on a mesh with a pointcloud behind it would cause the mouse event to target the pointcloud and not the mesh. This new logic treats pointclouds and meshes equally so that whatever is the closest object clicked is what the mouse event "snaps" to.